[plugins] Standardize plugin cfgs, draft v1
