+++ javalin/src/test/java/io/javalin/TestConfiguration.kt
@@ -39,9 +39,9 @@ class TestConfiguration {
             it.http.asyncTimeout = 10_000L
             it.http.generateEtags = true
             it.http.defaultContentType = ContentType.PLAIN
-            it.plugins.enableCors { cors -> cors.add { it.reflectClientOrigin = true } }
+            it.plugins.enableCors { cors -> cors.addRule { it.reflectClientOrigin = true } }
             it.plugins.enableDevLogging()
-            it.plugins.register(RouteOverviewPlugin("/test"))
+            it.plugins.enableRouteOverview("/test")
             it.plugins.enableSslRedirects()
             it.http.prefer405over404 = false
             it.requestLogger.http { ctx, timeInMs -> }
