+++ javalin/src/test/java/io/javalin/TestSslRedirectPlugin.kt
@@ -2,6 +2,7 @@ package io.javalin
 
 import io.javalin.http.Header
 import io.javalin.plugin.bundled.SslRedirectPlugin
+import io.javalin.plugin.bundled.SslRedirectPluginFactory
 import io.javalin.testing.TestUtil
 import kong.unirest.Unirest
 import org.assertj.core.api.Assertions.assertThat
@@ -17,8 +18,10 @@ class TestSslRedirectPlugin {
 
     @Test
     fun `ssl redirect plugin should redirect all http requests`() = TestUtil.test(
-        Javalin.create {
-            it.plugins.register(SslRedirectPlugin(redirectOnLocalhost = true))
+        Javalin.create { cfg ->
+            cfg.plugins.register(SslRedirectPluginFactory) {
+                it.redirectOnLocalhost = true
+            }
         }
     ) { app, http ->
         app.get("/") { ctx -> ctx.result("Hello") }
@@ -30,8 +33,11 @@ class TestSslRedirectPlugin {
 
     @Test
     fun `ssl redirect plugin should support custom ssl port`() = TestUtil.test(
-        Javalin.create {
-            it.plugins.register(SslRedirectPlugin(redirectOnLocalhost = true, sslPort = 8443))
+        Javalin.create { cfg ->
+            cfg.plugins.register(SslRedirectPluginFactory) {
+                it.redirectOnLocalhost = true
+                it.sslPort = 8443
+            }
         }
     ) { app, http ->
         app.get("/") { ctx -> ctx.result("Hello") }
