+++ javalin/src/test/java/io/javalin/routeoverview/TestRouteOverviewPlugin.kt
@@ -2,7 +2,6 @@ package io.javalin.routeoverview
 
 import io.javalin.Javalin
 import io.javalin.http.HandlerType
-import io.javalin.plugin.bundled.RouteOverviewPlugin
 import io.javalin.testing.TestUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -12,9 +11,7 @@ class TestRouteOverviewPlugin {
     @Test
     fun `should properly register and expose route overview`() = TestUtil.test(
         Javalin.create { config ->
-            config.plugins.register(
-                RouteOverviewPlugin("/overview")
-            )
+            config.plugins.enableRouteOverview(("/overview"))
         }
     ) { app, http ->
         VisualTest.setupJavalinRoutes(app)
